% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/encontrar_pior_ano_time.R
\name{encontrar_pior_ano_time}
\alias{encontrar_pior_ano_time}
\title{encontrar pior ano time}
\usage{
encontrar_pior_ano_time(time)
}
\arguments{
\item{time}{primeiro parâmetro}
}
\value{
descrição do resultado
}
\description{
serve para blabla
}
