% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{partidas_brasileirao}
\alias{partidas_brasileirao}
\title{Dados das partidas do Brasileirão}
\format{
Um data frame com colunas:
\describe{
  \item{coluna1}{Descrição}
}
}
\usage{
partidas_brasileirao
}
\description{
Dados das partidas do Brasileirão
}
\keyword{datasets}
